IDL: Put default values on 2 members of SecurityPolicyViolationEventInit
The dictionary in spec has a default value empty string on "referer" and
"sample", but we did not have.
This CL adds the default values, and updates some tests expectation.

Updated tests expected an event SecurityPolicyViolationEvent returns
null for "referer" and "sample" attributes, but their type are string
and shouldn't be null.  This CL updates to expect empty strings.

Bug: 839389
Change-Id: If8c9a9c2c7b2882ecbd0ab63df467e2b1bafa047